Move supplies claim to dungeon scene
