+++ client/Assets/AddressableAssetsData/OSX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58847661d284d4a10b0d6308bcd46a96
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
