+++ client/Assets/Scripts/Dungeon/DungeonManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a54eae7bcc7b1478087ea0ca47ab66aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
