+++ client/Assets/Scripts/Dungeon/DungeonSettlementManager.cs
@@ -15,7 +15,6 @@ public class DungeonSettlementManager : MonoBehaviour
     void Start()
     {
         GlobalUserData user = GlobalUserData.Instance;
-        ClaimRewards();
         SetSceneTexts(user.User);
     }
 
@@ -82,24 +81,6 @@ public class DungeonSettlementManager : MonoBehaviour
                 Debug.LogError($"Unhandled currency: {afkRewardRate.currency}");
             }
         }
-    }
-
-    private void ClaimRewards()
-    {
-        SocketConnection.Instance.ClaimDungeonAfkRewards(GlobalUserData.Instance.User.id, (userReceived) =>
-        {
-            GlobalUserData userToUpdate = GlobalUserData.Instance;
-            Dictionary<string, int> currenciesToAdd = new Dictionary<string, int>();
-
-            userReceived.currencies.Select(c => c.Key).ToList().ForEach(c =>
-            {
-                if (!currenciesToAdd.ContainsKey(c))
-                {
-                    currenciesToAdd.Add(c, userReceived.currencies[c] - userToUpdate.GetCurrency(c).Value);
-                }
-            });
-            userToUpdate.AddCurrencies(currenciesToAdd);
-        });
     }
 
     private string GetAfkRewardRateText(float daily_rate)
